de-couple search from search info
