+++ src/search/mod.rs
@@ -1,5 +1,6 @@
 mod definitions;
 mod eval;
-pub mod minimax;
 pub mod options;
 mod pv_table;
+pub mod search;
+mod timeman;
